backup/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
attrs==22.2.0
    # via pytest
autoflake==2.0.0
    # via -r requirements.in
boto3==1.24.59
    # via -r requirements.in
botocore==1.27.59
    # via
    #   boto3
    #   s3transfer
bson==0.5.10
    # via -r requirements.in
coverage[toml]==7.0.5
    # via pytest-cov
dnspython==2.2.1
    # via pymongo
dynaconf==3.1.11
    # via -r requirements.in
exceptiongroup==1.1.0
    # via pytest
ffprobe-python==1.0.3
    # via -r requirements.in
flake8==6.0.0
    # via -r requirements.in
flake8-fixme==1.1.1
    # via -r requirements.in
importlib-metadata==6.0.0
    # via pytest-randomly
iniconfig==2.0.0
    # via pytest
isort==5.11.4
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
loguru==0.6.0
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mimesis==4.1.3
    # via
    #   -r requirements.in
    #   pytest-mimesis
multidict==6.0.4
    # via pulsar
mypy==0.991
    # via -r requirements.in
mypy-extensions==0.4.3
    # via mypy
packaging==23.0
    # via pytest
pluggy==1.0.0
    # via pytest
pulsar==2.0.2
    # via -r requirements.in
pycodestyle==2.10.0
    # via flake8
pydantic==1.10.4
    # via -r requirements.in
pyflakes==3.0.1
    # via
    #   autoflake
    #   flake8
pymongo==4.3.3
    # via -r requirements.in
pytest==7.2.0
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-env
    #   pytest-mimesis
    #   pytest-randomly
pytest-asyncio==0.20.3
    # via -r requirements.in
pytest-cov==4.0.0
    # via -r requirements.in
pytest-env==0.8.1
    # via -r requirements.in
pytest-mimesis==1.1.0
    # via -r requirements.in
pytest-randomly==3.12.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   bson
python-dotenv==0.21.0
    # via -r requirements.in
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   bson
    #   python-dateutil
tomli==2.0.1
    # via
    #   autoflake
    #   coverage
    #   mypy
    #   pytest
typing-extensions==4.4.0
    # via
    #   mypy
    #   pydantic
urllib3==1.26.14
    # via botocore
watchdog==2.2.1
    # via -r requirements.in
zipp==3.11.0
    # via importlib-metadata
